% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_titer_lims.R
\name{calc_titer_diff_lims}
\alias{calc_titer_diff_lims}
\title{Calculate the mean difference between two paired sets of titers}
\usage{
calc_titer_diff_lims(titers1, titers2, dilution_stepsize, options = list())
}
\arguments{
\item{titers1}{The first titerset from which to calculate the mean difference}

\item{titers2}{The second titerset to which to calculate the mean difference}

\item{dilution_stepsize}{The dilution stepsize used in the assay, see `calc_titer_lims()`}

\item{options}{A named list of options to pass to `titer_fit_options()`}
}
\description{
This function is useful for example for calculating differences where you have a set of 
pre-vaccination and post-vaccination samples and you would like to know the mean response 
size, accounting for non-detectable values.
}
