% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_titer_diffs.R
\name{mean_titer_diffs}
\alias{mean_titer_diffs}
\title{Calculate the mean difference between two paired sets of titers}
\usage{
mean_titer_diffs(titers1, titers2, method, level = 0.95, dilution_stepsize)
}
\arguments{
\item{titers1}{The first titerset from which to calculate the mean difference}

\item{titers2}{The second titerset to which to calculate the mean difference}

\item{method}{The method to use when dealing with censored titers (like
`<10`), one of "replace_nd", "exclude_nd", "truncated_normal"}

\item{level}{The confidence level to use when calculating confidence intervals}

\item{dilution_stepsize}{The dilution stepsize used in the assay, see `calc_titer_lims()`}
}
\description{
This function is useful for example for calculating differences where you have a set of 
pre-vaccination and post-vaccination samples and you would like to know the mean response 
size, accounting for non-detectable values.
}
