% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_titers.R
\name{mean_titers}
\alias{mean_titers}
\title{Calculate the geometric mean titer of a set of titers}
\usage{
mean_titers(
  titers,
  method,
  level = 0.95,
  sd = NA,
  dilution_stepsize,
  options = list()
)
}
\arguments{
\item{titers}{A vector of titers}

\item{method}{The method to use when dealing with censored titers (like
`<10`), one of "replace_nd", "exclude_nd", "truncated_normal"}

\item{level}{The confidence level to use when calculating confidence intervals}

\item{sd}{By setting this you can fix the standard deviation assumed when finding 
parameters for the normal distribution when using the "truncated_normal" approach}

\item{dilution_stepsize}{The dilution stepsize used in the assay, see `calc_titer_lims()`}

\item{options}{A named list of options to pass to `titer_fit_options()`}
}
\description{
Calculate the geometric mean titer of a set of titers
}
