\name{meantiter-package}
\alias{meantiter-package}
\alias{meantiter}
\docType{package}
\title{
\packageTitle{meantiter}
}
\description{
\packageDescription{meantiter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{meantiter}
\packageIndices{meantiter}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{meantiter}

Maintainer: \packageMaintainer{meantiter}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
